Updated timestamp calculation

+++ Bluepill-runner/Bluepill-runner/BPTraceEvent.h
@@ -35,7 +35,7 @@
 - (void)appendCompleteTraceEvent:(NSString *)name
                                 :(NSString *)category
                                 :(NSInteger)timestamp
-                                :(float)duration
+                                :(NSInteger)duration
                                 :(NSInteger)process_id
                                 :(NSInteger)thread_id
                                 :(NSDictionary *)args;
